fix(build): restore the archive-OFF CI gate; add me-archive's defence-in-depth floor
Addresses the adversarial review of the me-archive carve. Two real findings, neither
caught by any green gate.

1. The carve silently killed the archive-OFF CI gate.

me-archive is the first workspace member to enable a sibling's optional feature
UNCONDITIONALLY (me-types/me-storage/me-index with features = ["archive"]); every other
crate declares its own optional `archive` feature and forwards. That is correct for
me-archive — it IS the archive primitive and cannot compile without them — but because it
is an unconditional workspace member, Cargo unifies features across a --workspace build.
So those three crates now compile with `archive` ON in EVERY workspace job, including
`cargo build --workspace` (default features).

Proven, not deduced: `cargo tree --workspace -e features -i me-types` reports
['default','test-util'] at base and ['archive','default','test-util'] at HEAD, with
me-archive as the sole enabler.

Net effect: no CI job compiled me-types/me-storage/me-index archive-OFF any more — the
exact configuration a downstream consumer of `memory-engine` resolves to by default. A
future PR could break their archive-off compile and pass fmt, clippy, both builds and the
whole test suite. This is the "--all-features masks single-feature regressions" trap,
extended to the *default* workspace build.

Adds facade-alone `cargo check -p memory-engine` (default and --no-default-features) to
the CI Build job: `-p` narrows the graph to exclude me-archive and restores the gate. Both
pass today. Documented in me-archive/Cargo.toml so they are not later deleted as
"redundant with --workspace" — they are not.

2. me_archive::archive enforced neither guard MemoryCtx hands it.

It is a `pub`, destructive primitive (writes a .pak, hard-deletes facts and edges) and, in
base, `ensure_open()` lived inside the sole entry point. The carve turned it into a caller
obligation honoured only by the facade delegate.

Adds `ctx.ensure_open()?`, matching every other carved primitive (me-query, me-ingest,
me-resolve all open with it). Behaviour-neutral — the delegate's identical check fires
first, so no input changes outcome or error order — but the #742 reopen fence has no
backend-level enforcement: a fenced handle (post different-dim reconstruction) would
otherwise stamp the STALE ctx.embed_dim into a .pak whose facts carry D'-width embeddings.
Archive search then silently drops embedding scoring for that pak: permanent, silent
cold-storage corruption.

`ctx.ensure_writable()` is deliberately NOT added: no carved primitive calls it, that gate
is tracked systemically across all write primitives in #972, and read-only is caught below
the seam anyway (block_write -> try_write -> MemoryError::ReadOnly). Adding it here alone
would pre-empt #972's decision.

Also: discloses a guard-order change in `verify_archives` (archive_dir now resolves before
the manifest read — a fail-fast improvement, but the carve claimed faithfulness, so it is
recorded); and relocates `pak_write_stamp_matches_read_gate` into me-archive, where the
property now lives. That test previously stood up an entire SQLite engine purely to read
back its embed_dim; as a me-archive unit test it needs no engine, which lets `build_pak`
narrow from `pub` to `pub(crate)`.

anti-#903: 1970 passed / 0 failed (the test moved; nothing added or dropped).

Refs #940
